refactor: reorganizar e atualizar tipos de moderação
- Criado novo arquivo para ModerationTargetType
- Atualizados imports para utilizar o novo tipo
- Ajustados tipos em CreateReportDto e serviços relacionados
