+++ central-pet-back/src/modules/moderation/moderation-target-type.ts
@@ -1,0 +1,7 @@
+export const ModerationTargetType = {
+  PET: 'PET',
+  USER: 'USER',
+  ADOPTION_REQUEST: 'ADOPTION_REQUEST',
+} as const;
+
+export type ModerationTargetType = (typeof ModerationTargetType)[keyof typeof ModerationTargetType];
